requirements.txt
fastapi
uvicorn
pandas
networkx
mysql-connector-python
python-multipart